- Runtime tests - small reorganization. - Added some negative cases.
