+++ kie-dmn-core/src/test/java/org/kie/dmn/core/DMNRuntimeTest.java
@@ -16,162 +16,43 @@
 
 package org.kie.dmn.core;
 
-import org.junit.Ignore;
-import org.junit.Test;
-import org.kie.api.KieServices;
-import org.kie.api.runtime.KieContainer;
-import org.kie.dmn.core.api.*;
-import org.kie.dmn.core.api.event.*;
-import org.kie.dmn.core.ast.InputDataNode;
-import org.mockito.ArgumentCaptor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.util.*;
-
 import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.CoreMatchers.nullValue;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.hamcrest.Matchers.hasEntry;
+import static org.junit.Assert.*;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.internal.verification.VerificationModeFactory.times;
 
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import org.junit.Ignore;
+import org.junit.Test;
+import org.kie.dmn.core.api.DMNContext;
+import org.kie.dmn.core.api.DMNFactory;
+import org.kie.dmn.core.api.DMNMessage;
+import org.kie.dmn.core.api.DMNModel;
+import org.kie.dmn.core.api.DMNResult;
+import org.kie.dmn.core.api.DMNRuntime;
+import org.kie.dmn.core.api.event.AfterEvaluateDecisionEvent;
+import org.kie.dmn.core.api.event.AfterEvaluateDecisionTableEvent;
+import org.kie.dmn.core.api.event.BeforeEvaluateDecisionEvent;
+import org.kie.dmn.core.api.event.BeforeEvaluateDecisionTableEvent;
+import org.kie.dmn.core.api.event.DMNRuntimeEventListener;
+import org.kie.dmn.core.util.DMNRuntimeUtil;
+import org.mockito.ArgumentCaptor;
+
 public class DMNRuntimeTest {
-
-    protected DMNRuntime createRuntime( String resourceName ) {
-        KieServices ks = KieServices.Factory.get();
-        KieContainer kieContainer = KieHelper.getKieContainer(
-                ks.newReleaseId( "org.kie", "dmn-test", "1.0" ),
-                ks.getResources().newClassPathResource( resourceName, DMNRuntimeTest.class ) );
-
-        DMNRuntime runtime = kieContainer.newKieSession().getKieRuntime( DMNRuntime.class );
-        assertNotNull( runtime );
-        return runtime;
-    }
-
-    @Test
-    public void testSimpleEvaluateAll() {
-        DMNRuntime runtime = createRuntime( "0001-input-data-string.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0001-input-data-string" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Full Name", "John Doe" );
-
-        DMNResult dmnResult = runtime.evaluateAll( dmnModel, context );
-
-        assertThat( dmnResult.getDecisionResults().size(), is(1) );
-        assertThat( dmnResult.getDecisionResultByName( "Greeting Message" ).getResult(), is( "Hello John Doe" ) );
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat( result.get( "Greeting Message" ), is( "Hello John Doe" ) );
-    }
-
-    @Test
-    public void testSimpleEvaluateDecisionByName() {
-        DMNRuntime runtime = createRuntime( "0001-input-data-string.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0001-input-data-string" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Full Name", "John Doe" );
-
-        DMNResult dmnResult = runtime.evaluateDecisionByName( dmnModel, "Greeting Message", context );
-
-        assertThat( dmnResult.getDecisionResults().size(), is(1) );
-        assertThat( dmnResult.getDecisionResultByName( "Greeting Message" ).getResult(), is( "Hello John Doe" ) );
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat( result.get( "Greeting Message" ), is( "Hello John Doe" ) );
-    }
-
-    @Test
-    public void testSimpleEvaluateDecisionById() {
-        DMNRuntime runtime = createRuntime( "0001-input-data-string.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0001-input-data-string" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Full Name", "John Doe" );
-
-        DMNResult dmnResult = runtime.evaluateDecisionById( dmnModel, "d_GreetingMessage", context );
-
-        assertThat( dmnResult.getDecisionResults().size(), is(1) );
-        assertThat( dmnResult.getDecisionResultById( "d_GreetingMessage" ).getResult(), is( "Hello John Doe" ) );
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat( result.get( "Greeting Message" ), is( "Hello John Doe" ) );
-    }
-
-    @Test
-    public void testGetRequiredInputsByName() {
-        DMNRuntime runtime = createRuntime( "0001-input-data-string.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0001-input-data-string" );
-        assertThat( dmnModel, notNullValue() );
-
-        Set<InputDataNode> inputs = dmnModel.getRequiredInputsForDecisionName( "Greeting Message" );
-
-        assertThat( inputs.size(), is(1) );
-        assertThat( inputs.iterator().next().getName(), is("Full Name") );
-    }
-
-    @Test
-    public void testGetRequiredInputsById() {
-        DMNRuntime runtime = createRuntime( "0001-input-data-string.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0001-input-data-string" );
-        assertThat( dmnModel, notNullValue() );
-
-        Set<InputDataNode> inputs = dmnModel.getRequiredInputsForDecisionId( "d_GreetingMessage" );
-
-        assertThat( inputs.size(), is(1) );
-        assertThat( inputs.iterator().next().getName(), is("Full Name") );
-    }
-
-    @Test
-    public void testSimpleInputDataNumber() {
-        DMNRuntime runtime = createRuntime( "0002-input-data-number.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0002-input-data-number" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Monthly Salary", new BigDecimal( 1000 ) );
-
-        DMNResult dmnResult = runtime.evaluateAll( dmnModel, context );
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat( result.get( "Yearly Salary" ), is( new BigDecimal( 12000 ) ) );
-    }
-
-    @Test
-    public void testSimpleAllowedValuesString() {
-        DMNRuntime runtime = createRuntime( "0003-input-data-string-allowed-values.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0003-input-data-string-allowed-values" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Employment Status", "SELF-EMPLOYED" );
-
-        DMNResult dmnResult = runtime.evaluateAll( dmnModel, context );
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat( result.get( "Employment Status Statement" ), is( "You are SELF-EMPLOYED" ) );
-    }
 
     @Test
     public void testCompositeItemDefinition() {
-        DMNRuntime runtime = createRuntime( "0008-LX-arithmetic.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "0008-LX-arithmetic.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0008-LX-arithmetic" );
         assertThat( dmnModel, notNullValue() );
 
@@ -190,45 +71,8 @@ public class DMNRuntimeTest {
     }
 
     @Test
-    public void testSimpleDTUnique() {
-        DMNRuntime runtime = createRuntime( "0004-simpletable-U.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0004-simpletable-U" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Age", new BigDecimal( 18 ) );
-        context.set( "RiskCategory", "Medium" );
-        context.set( "isAffordable", true );
-
-        DMNResult dmnResult = runtime.evaluateAll( dmnModel, context );
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat( result.get( "Approval Status" ), is( "Approved" ) );
-    }
-    
-    @Test
-    public void testSimpleDTUniqueSatisfies() {
-        DMNRuntime runtime = createRuntime( "0004-simpletable-U.dmn" );
-        DMNModel dmnModel = runtime.getModel( "https://github.com/droolsjbpm/kie-dmn", "0004-simpletable-U" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Age", new BigDecimal( 18 ) );
-        context.set( "RiskCategory", "ASD" );
-        context.set( "isAffordable", false );
-
-        DMNResult dmnResult = runtime.evaluateAll( dmnModel, context );
-
-        DMNContext result = dmnResult.getContext();
-
-        assertThat( result.get( "Approval Status" ), nullValue() );
-        assertTrue( dmnResult.getMessages().size() > 0 ); 
-    }
-
-    @Test
     public void testTrisotechNamespace() {
-        DMNRuntime runtime = createRuntime( "trisotech_namespace.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "trisotech_namespace.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/dmn/definitions/_b8feec86-dadf-4051-9feb-8e6093bbb530", "Solution 3" );
         assertThat( dmnModel, notNullValue() );
 
@@ -243,7 +87,7 @@ public class DMNRuntimeTest {
 
     @Test
     public void testEmptyDecision1() {
-        DMNRuntime runtime = createRuntime( "empty_decision.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "empty_decision.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/dmn/definitions/_ba9fc4b1-5ced-4d00-9b61-290de4bf3213", "Solution 3" );
         assertThat( dmnModel, notNullValue() );
 
@@ -264,7 +108,7 @@ public class DMNRuntimeTest {
 
     @Test
     public void testEmptyDecision2() {
-        DMNRuntime runtime = createRuntime( "empty_decision.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "empty_decision.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/dmn/definitions/_ba9fc4b1-5ced-4d00-9b61-290de4bf3213", "Solution 3" );
         assertThat( dmnModel, notNullValue() );
 
@@ -292,50 +136,12 @@ public class DMNRuntimeTest {
     }
 
     @Test
-    public void testDecisionTableWithCalculatedResult() {
-        DMNRuntime runtime = createRuntime( "calculation1.dmn" );
-        DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/definitions/_77ae284e-ce52-4579-a50f-f3cc584d7f4b", "Calculation1" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "MonthlyDeptPmt", BigDecimal.valueOf( 200 ) );
-        context.set( "MonthlyPmt", BigDecimal.valueOf( 100 ) );
-        context.set( "MonthlyIncome", BigDecimal.valueOf( 600 ) );
-
-        DMNResult dmnResult = runtime.evaluateAll(dmnModel, context);
-        assertThat( dmnResult.hasErrors(), is( false ) );
-
-        DMNContext result = dmnResult.getContext();
-        assertThat( ((BigDecimal)result.get("Logique de décision 1")).setScale( 1, RoundingMode.CEILING), is( BigDecimal.valueOf( 0.5 ) ) );
-    }
-
-    @Test
-    public void testDecisionTableMultipleResults() {
-        DMNRuntime runtime = createRuntime( "car_damage_responsibility.dmn" );
-        DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/definitions/_820611e9-c21c-47cd-8e52-5cba2be9f9cc", "Car Damage Responsibility" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Membership Level", "Silver" );
-        context.set( "Damage Types", "Body" );
-        context.set( "Responsible", "Driver" );
-
-        DMNResult dmnResult = runtime.evaluateAll(dmnModel, context);
-        assertThat( dmnResult.hasErrors(), is( false ) );
-
-        DMNContext result = dmnResult.getContext();
-        assertThat( (Map<String,Object>)result.get("Car Damage Responsibility"), hasEntry( is( "EU Rent" ), is( BigDecimal.valueOf( 40 )) ));
-        assertThat( (Map<String,Object>)result.get("Car Damage Responsibility"), hasEntry( is( "Renter" ), is( BigDecimal.valueOf( 60 )) ));
-        assertThat( result.get("Payment method"), is( "Check" ) );
-    }
-
-    @Test
     public void testEventListeners() {
-        DMNRuntime runtime = createRuntime( "car_damage_responsibility.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "car_damage_responsibility.dmn", this.getClass() );
 
         DMNRuntimeEventListener listener = mock( DMNRuntimeEventListener.class );
         runtime.addListener( listener );
-        runtime.addListener( createListener() );
+        runtime.addListener( DMNRuntimeUtil.createListener() );
 
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/definitions/_820611e9-c21c-47cd-8e52-5cba2be9f9cc", "Car Damage Responsibility" );
         assertThat( dmnModel, notNullValue() );
@@ -374,59 +180,8 @@ public class DMNRuntimeTest {
     }
 
     @Test
-    public void testDecisionTableU() {
-        DMNRuntime runtime = createRuntime( "BranchDistribution.dmn" );
-        DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/dmn/definitions/_cdf29af2-959b-4004-8271-82a9f5a62147", "Dessin 1" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Branches dispersion", "Province" );
-        context.set( "Number of Branches", BigDecimal.valueOf( 10 ) );
-
-        DMNResult dmnResult = runtime.evaluateAll(dmnModel, context);
-        assertThat( dmnResult.hasErrors(), is( false ) );
-
-        DMNContext result = dmnResult.getContext();
-        System.out.println(result);
-        assertThat( result.get("Branches distribution"), is( "Medium" ) );
-    }
-
-    @Test
-    public void testInvalidInputDTErrorMessage() {
-        DMNRuntime runtime = createRuntime( "InvalidInput.dmn" );
-        DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/dmn/definitions/_cdf29af2-959b-4004-8271-82a9f5a62147", "Dessin 1" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Branches dispersion", "Province" );
-        context.set( "Number of Branches", BigDecimal.valueOf( 10 ) );
-
-        DMNResult dmnResult = runtime.evaluateAll(dmnModel, context);
-        assertThat( dmnResult.hasErrors(), is( true ) );
-
-        DMNContext result = dmnResult.getContext();
-        assertThat( result.isDefined("Branches distribution"), is( false ) );
-    }
-
-    @Test
-    public void testCollectHitPolicy() {
-        DMNRuntime runtime = createRuntime( "Collect_Hit_Policy.dmn" );
-        DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/definitions/_da1a4dcb-01bf-4dee-9be8-f498bc68178c", "Collect Hit Policy" );
-        assertThat( dmnModel, notNullValue() );
-
-        DMNContext context = DMNFactory.newContext();
-        context.set( "Input", 20 );
-
-        DMNResult dmnResult = runtime.evaluateAll(dmnModel, context);
-        assertThat( dmnResult.hasErrors(), is( false ) );
-
-        DMNContext result = dmnResult.getContext();
-        assertThat( result.get("Collect"), is( BigDecimal.valueOf( 50 ) ) );
-    }
-
-    @Test
     public void testOutputReuse() {
-        DMNRuntime runtime = createRuntime( "Input_reuse_in_output.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "Input_reuse_in_output.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/dmn/definitions/_098bb607-eff7-4772-83ac-6ded8b371fa7", "Input reuse in output" );
         assertThat( dmnModel, notNullValue() );
 
@@ -444,7 +199,7 @@ public class DMNRuntimeTest {
 
     @Test
     public void testSimpleNot() {
-        DMNRuntime runtime = createRuntime( "Simple_Not.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "Simple_Not.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/definitions/_98436ebb-7c42-48c0-8d11-d693e2a817c9", "Simple Not" );
         assertThat( dmnModel, notNullValue() );
 
@@ -461,7 +216,7 @@ public class DMNRuntimeTest {
 
     @Test
     public void testSimpleNot2() {
-        DMNRuntime runtime = createRuntime( "Simple_Not.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "Simple_Not.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/definitions/_98436ebb-7c42-48c0-8d11-d693e2a817c9", "Simple Not" );
         assertThat( dmnModel, notNullValue() );
 
@@ -478,7 +233,7 @@ public class DMNRuntimeTest {
 
     @Test
     public void testDinner() {
-        DMNRuntime runtime = createRuntime( "Dinner.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "Dinner.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/definitions/_0c45df24-0d57-4acc-b296-b4cba8b71a36", "Dinner" );
         assertThat( dmnModel, notNullValue() );
         assertThat( dmnModel.hasErrors(), is(false) );
@@ -499,7 +254,7 @@ public class DMNRuntimeTest {
 
     @Test @Ignore("not implemented yet")
     public void testBoxedContext() {
-        DMNRuntime runtime = createRuntime( "BoxedContext.dmn" );
+        DMNRuntime runtime = DMNRuntimeUtil.createRuntime( "BoxedContext.dmn", this.getClass() );
         DMNModel dmnModel = runtime.getModel( "http://www.trisotech.com/dmn/definitions/_0de36357-fec0-4b4e-b7f1-382d381e06e9", "Dessin 1" );
         assertThat( dmnModel, notNullValue() );
         assertThat( dmnModel.getMessages().toString(), dmnModel.hasErrors(), is(false) );
@@ -512,31 +267,5 @@ public class DMNRuntimeTest {
         assertThat( dmnResult.hasErrors(), is( false ) );
         assertThat( (Map<String, Object>) dmnResult.getContext().get( "Math" ), hasEntry( "Sum", BigDecimal.valueOf( 15 ) ) );
         assertThat( (Map<String, Object>) dmnResult.getContext().get( "Math" ), hasEntry( "Product", BigDecimal.valueOf( 50 ) ) );
-    }
-
-    private DMNRuntimeEventListener createListener() {
-        return new DMNRuntimeEventListener() {
-            private final Logger logger = LoggerFactory.getLogger( DMNRuntimeEventListener.class );
-
-            @Override
-            public void beforeEvaluateDecision(BeforeEvaluateDecisionEvent event) {
-                logger.info( event.toString() );
-            }
-
-            @Override
-            public void afterEvaluateDecision(AfterEvaluateDecisionEvent event) {
-                logger.info( event.toString() );
-            }
-
-            @Override
-            public void beforeEvaluateDecisionTable(BeforeEvaluateDecisionTableEvent event) {
-                logger.info( event.toString() );
-            }
-
-            @Override
-            public void afterEvaluateDecisionTable(AfterEvaluateDecisionTableEvent event) {
-                logger.info( event.toString() );
-            }
-        };
     }
 }
