+++ kie-dmn-core/src/test/java/org/kie/dmn/core/util/DMNRuntimeUtil.java
@@ -1,0 +1,74 @@
+/*
+ * Copyright 2016 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.dmn.core.util;
+
+import org.junit.Assert;
+import org.kie.api.KieServices;
+import org.kie.api.runtime.KieContainer;
+import org.kie.dmn.core.KieHelper;
+import org.kie.dmn.core.api.DMNRuntime;
+import org.kie.dmn.core.api.event.AfterEvaluateDecisionEvent;
+import org.kie.dmn.core.api.event.AfterEvaluateDecisionTableEvent;
+import org.kie.dmn.core.api.event.BeforeEvaluateDecisionEvent;
+import org.kie.dmn.core.api.event.BeforeEvaluateDecisionTableEvent;
+import org.kie.dmn.core.api.event.DMNRuntimeEventListener;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public final class DMNRuntimeUtil {
+
+    public static DMNRuntime createRuntime(final String resourceName, final Class testClass) {
+        final KieServices ks = KieServices.Factory.get();
+        final KieContainer kieContainer = KieHelper.getKieContainer(
+                ks.newReleaseId("org.kie", "dmn-test", "1.0"),
+                ks.getResources().newClassPathResource(resourceName, testClass));
+
+        final DMNRuntime runtime = kieContainer.newKieSession().getKieRuntime(DMNRuntime.class);
+        Assert.assertNotNull(runtime);
+        return runtime;
+    }
+
+    public static DMNRuntimeEventListener createListener() {
+        return new DMNRuntimeEventListener() {
+            private final Logger logger = LoggerFactory.getLogger(DMNRuntimeEventListener.class);
+
+            @Override
+            public void beforeEvaluateDecision(BeforeEvaluateDecisionEvent event) {
+                logger.info(event.toString());
+            }
+
+            @Override
+            public void afterEvaluateDecision(AfterEvaluateDecisionEvent event) {
+                logger.info(event.toString());
+            }
+
+            @Override
+            public void beforeEvaluateDecisionTable(BeforeEvaluateDecisionTableEvent event) {
+                logger.info(event.toString());
+            }
+
+            @Override
+            public void afterEvaluateDecisionTable(AfterEvaluateDecisionTableEvent event) {
+                logger.info(event.toString());
+            }
+        };
+    }
+
+    private DMNRuntimeUtil() {
+        // No constructor for util class.
+    }
+}
